# optimization: reduce GPU transfers