This PR adds a `ParamCache` to the qparams file, which manages a cache of `nn.Parameter(option[idx].weight.cuda())` elements.

It also adds a new `batch_test_error` function which takes a model factory and factory parameters, and processes each layer in turn, for all model variations, i.e. the opposite of what is happening already.

This gives a 30% increase in speed for the attention layer, at a cost of some additional VRAM usage.

However, applying this method to the `measure_mlp` function does not give significant speed-ups, so the VRAM bump is not worth it. I'm still investigating why this is the case.

Supercedes (incorporates) #455.

This PR primarily targets the calibration feature, but the .item() tweak to test_error snuck its way into the quantize part. Can remove. Do intend to address quantization optimizations.

```
==================================================================

                               attn

==================================================================

BASE MASTER:

VRAM PEAK = 2715 MB

--------------------------------------------
| Measured: model.layers.0 (Attention)     |
| Duration: 96.49 seconds                  |
| Completed step: 1/163                    |
| Avg time / step (rolling): 96.49 seconds |
| Estimated remaining time: 260min 31sec   |
| Last checkpoint layer: None              |
--------------------------------------------

------------------------------------------------------------------

MASTER + PARAMCACHE + BATCHING

VRAM PEAK = 6200 MB

--------------------------------------------
| Measured: model.layers.0 (Attention)     |
| Duration: 60.14 seconds                  |
| Completed step: 1/163                    |
| Avg time / step (rolling): 60.14 seconds |
| Estimated remaining time: 162min 23sec   |
| Last checkpoint layer: None              |
--------------------------------------------
```