@kallewoof 2024-05-16
> I did at one point keep the entire calibration state in VRAM, which was significantly faster, but it also made it impossible to quantize large models. Then I made it switchable, causing great confusion, and finally I arrived at always swapping as the best compromise between reliability and performance.

I think this modification results in zero peak VRAM increase (because the MLP layer processing uses more even without the batching), and is a pretty significant boost to the attn layer processing speed.

> This is also in part because, as you've noted, PCIe traffic matters much less when processing the MLP modules, and they account for most of the time spent in measurement. It's a real struggle to get a large enough chunk of the full-precision model to fit sometimes.

I see! Thanks for confirming that.

> One thing I've considered is a better pipeline for swapping. Currently all data is swapped synchronously, but it could be asynchronous via a pinned buffer. It could maybe work in conjunction with this caching mechanism.

Not sure how that works but will look around.

> I'm currently a little preoccupied with dynamic batching, but I'll give it all some attention real soon.

Thanks for the heads up, I'll be patient. :)