@kallewoof 2024-05-22
I'm closing this as I'm personally not convinced myself that it's worth it with the added technical debt. A simple 'cache hidden/training states up to a given amount' would probably be good though, and simpler, so I'm going to take a stab at that.